+++ tablet_qt/lib/uifunc.cpp
@@ -608,7 +608,7 @@ QString escapeString(const QString& string)
     const QByteArray arr = string.toLatin1();
     const int len = arr.length();
     QString result;
-    result.reserve(len * 1.1);  // as per QString::toHtmlEscaped
+    result.reserve(static_cast<int>(len * 1.1));  // as per QString::toHtmlEscaped
     result.append('"');  // opening quote
     for (int i = 0; i < len; ++i) {
         char c = arr.at(i);
